backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMySQL==1.1.0
python-dotenv==1.0.0
Pillow==12.0.0
djangorestframework-simplejwt==5.3.0
django-filter==23.2
